Update layout.tsx
Removed unnecessary import statement.

And certain improvements also such as providing type annotations for the props of the 'Layout'.

Instead of using the html and head tags manually, you can use the Head component from next/head to manage the head section of your document.

Removed the unnecessary 'null' check when returning the session from the 'getSession' function. The function can return 'session' directly, and if it's an empty object, it will evaluate to 'false'.

Used object destructuring for 'props' in the 'Layout' component's parameter to make your code more consistent and avoid confusion.